+++ src/components/common/icons/index.ts
@@ -1,4 +1,4 @@
-export { default as AccountCircleIcon } from "./AccountCircleIcon";
 export { default as DarkModeIcon } from "./DarkModeIcon";
 export { default as LightModeIcon } from "./LightModeIcon";
+export { default as LogoutIcon } from "./LogoutIcon";
 export { default as MenuIcon } from "./MenuIcon";
